处理11.3升级出现的警告

+++ SJVideoPlayer/ControlView/SJVideoPlayerFilmEditingControlView.m
@@ -48,7 +48,7 @@ NS_ASSUME_NONNULL_END
 
 - (void)dealloc {
 #ifdef DEBUG
-    NSLog(@"SJVideoPlayerLog: %zd - %s", __LINE__, __func__);
+    NSLog(@"SJVideoPlayerLog: %d - %s", (int)__LINE__, __func__);
 #endif
 }
