+++ SJVideoPlayer/SJVideoPlayerDefaultControlView.m
@@ -100,7 +100,7 @@ NS_ASSUME_NONNULL_END
 
 - (void)dealloc {
 #ifdef DEBUG
-    NSLog(@"SJVideoPlayerLog: %zd - %s", __LINE__, __func__);
+    NSLog(@"SJVideoPlayerLog: %d - %s", (int)__LINE__, __func__);
 #endif
     [[NSNotificationCenter defaultCenter] removeObserver:self];
 }
